+++ spec/matchers/aspect_ratio_validator_matcher_spec.rb
@@ -309,4 +309,8 @@ RSpec.describe ActiveStorageValidations::Matchers::AspectRatioValidatorMatcher d
   describe "#except_on" do
     it_behaves_like "works with except_on"
   end
+
+  describe "#timeout" do
+    it_behaves_like "works with timeout"
+  end
 end