+++ spec/matchers/content_type_validator_matcher_spec.rb
@@ -226,6 +226,10 @@ RSpec.describe ActiveStorageValidations::Matchers::ContentTypeValidatorMatcher d
     it_behaves_like "works with except_on"
   end
 
+  describe "#timeout" do
+    it_behaves_like "works with timeout"
+  end
+
   describe "Combinations" do
     describe "#allowing + #with_message" do
       let(:model_attribute) { :allowing_one_with_message }