+++ spec/matchers/dimension_validator_matcher_spec.rb
@@ -309,6 +309,10 @@ RSpec.describe ActiveStorageValidations::Matchers::DimensionValidatorMatcher do
     it_behaves_like "works with except_on"
   end
 
+  describe "#timeout" do
+    it_behaves_like "works with timeout"
+  end
+
   describe "Combinations" do
     %i[width height].each do |dimension|
       describe "##{dimension} + #with_message" do